User module refactoring.
