+++ src/common/examples/open-api-examples.ts
