+++ src/common/helpers/decorators.ts
@@ -1,0 +1,3 @@
+import { Param, ParseUUIDPipe } from '@nestjs/common';
+
+export const UUIDParam = (name: string) => Param(name, new ParseUUIDPipe());
