+++ src/user/dto/create-user-dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
-import { ErrorMessages } from 'src/constants/error-messages';
+import { ErrorMessages } from 'src/common/constants/error-messages';
 
 export class CreateUserDto {
   @ApiProperty({ example: 'user_4', description: 'User login' })
