--- src/user/params/get-user-by-id-params.ts
@@ -1,3 +1,0 @@
-export class GetUserByIdParams {
-  readonly id: string;
-}
